refactor test method according to interface change of KSVD #24
