+++ spmimage/decomposition/dict_learning.py
@@ -1,10 +1,9 @@
 import numpy as np
 
 from sklearn.decomposition import sparse_encode
-from sklearn.externals.joblib import Parallel, delayed
 
 
-def sparse_encode_with_mask(X, dictionary, mask, n_jobs=1, verbose=False, **kwargs):
+def sparse_encode_with_mask(X, dictionary, mask, **kwargs):
     """sparse_encode_with_mask
     Finds a sparse coding that represent data with given dictionary.
 
@@ -42,12 +41,8 @@ def sparse_encode_with_mask(X, dictionary, mask, n_jobs=1, verbose=False, **kwar
             The sparse codes
     """
     code = np.zeros((X.shape[0], dictionary.shape[0]))
-    codes = Parallel(n_jobs=n_jobs, verbose=verbose)(
-        delayed(sparse_encode)(
-            X[idx, :][mask[idx, :] == 1].reshape(1, -1),
-            dictionary[:, mask[idx, :] == 1],
-            **kwargs
-        ) for idx in range(X.shape[0]))
-    for idx, c in zip(range(X.shape[0]), codes):
-        code[idx, :] = c
+    for idx in range(X.shape[0]):
+        code[idx, :] = sparse_encode(X[idx, :][mask[idx, :] == 1].reshape(1, -1),
+                                     dictionary[:, mask[idx, :] == 1],
+                                     **kwargs)
     return code
